refactor text.js
