+++ .eslintrc.js
@@ -96,11 +96,11 @@ module.exports = {
         "max-depth": "error",
         "max-len": "off",
         "max-lines": "error",
-        "max-lines-per-function": "error",
+        "max-lines-per-function": "off",
         "max-nested-callbacks": "error",
         "max-params": "error",
-        "max-statements": "error",
-        "max-statements-per-line": "error",
+        "max-statements": "off",
+        "max-statements-per-line": "off",
         "multiline-comment-style": "error",
         "multiline-ternary": "error",
         "new-cap": "error",
