+++ src/robots/test/breakContentIntoSentences.js
@@ -1,0 +1,12 @@
+const sentenceBoundaryDetection = require('sbd')
+
+module.exports = 
+    function breakContentIntoSentences(sourceContentSanitize){
+		return sentenceBoundaryDetection
+			.sentences(sourceContentSanitize)
+			.map(sentence => ({
+				images: [],
+				keywords: [],
+				text: sentence
+			}))
+	}
